ENT-3140: Set the truststore for RHSM
The certificate used by the RHSM API endpoints is signed by a cert not
in the default system truststore. Therefore, we need to add it
explicitly to the truststore. Happily, our secret already has the proper
truststore set up, so only a template change is needed.

This is why neither I nor @mstead noticed this issue locally (we both
have system truststores with the proper certificate already in place).